tile-background/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anyio=3.7.1=pyhd8ed1ab_0
ca-certificates=2023.11.17=hbcca054_0
click=8.1.7=unix_pyh707e725_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
fastapi=0.108.0=pyhd8ed1ab_0
h11=0.14.0=pyhd8ed1ab_0
idna=3.6=pyhd8ed1ab_0
imageio=2.34.0=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
openssl=3.0.12=h7f8727e_0
packaging=24.0=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py39h06a4308_0
pydantic=1.10.12=py39h5eee18b_1
python=3.9.18=h955ad1f_0
python_abi=3.9=2_cp39
pywavelets=1.6.0=pypi_0
readline=8.2=h5eee18b_0
scikit-image=0.22.0=pypi_0
scipy=1.13.0=pypi_0
setuptools=68.2.2=py39h06a4308_0
sniffio=1.3.0=pyhd8ed1ab_0
sqlite=3.41.2=h5eee18b_0
starlette=0.32.0.post1=pyhd8ed1ab_0
tifffile=2024.2.12=pypi_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.9.0=hd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
uvicorn=0.25.0=py39hf3d152e_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.5=h5eee18b_0
zlib=1.2.13=h5eee18b_0
